Move hub classes into different name space based on what they're used for
