+++ osu.Server.Spectator.Tests/Multiplayer/MatchTypeTests.cs
@@ -10,7 +10,7 @@ using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Multiplayer.MatchTypes.TeamVersus;
 using osu.Game.Online.Rooms;
 using osu.Server.Spectator.Database.Models;
-using osu.Server.Spectator.Hubs;
+using osu.Server.Spectator.Hubs.Multiplayer;
 using Xunit;
 
 namespace osu.Server.Spectator.Tests.Multiplayer
