+++ osu.Server.Spectator.Tests/Multiplayer/MultiplayerAllPlayersQueueTests.cs
@@ -7,7 +7,7 @@ using Moq;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Server.Spectator.Database.Models;
-using osu.Server.Spectator.Hubs;
+using osu.Server.Spectator.Hubs.Multiplayer;
 using Xunit;
 
 namespace osu.Server.Spectator.Tests.Multiplayer
