+++ osu.Server.Spectator.Tests/Multiplayer/MultiplayerQueueTests.cs
@@ -9,7 +9,7 @@ using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Game.Rulesets;
 using osu.Server.Spectator.Database.Models;
-using osu.Server.Spectator.Hubs;
+using osu.Server.Spectator.Hubs.Multiplayer;
 using Xunit;
 
 namespace osu.Server.Spectator.Tests.Multiplayer
