+++ osu.Server.Spectator/Hubs/Metadata/MetadataHub.cs
@@ -1,0 +1,25 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Threading.Tasks;
+using osu.Game.Online.Metadata;
+using osu.Server.Spectator.Database;
+
+namespace osu.Server.Spectator.Hubs.Metadata
+{
+    public class MetadataHub : LoggingHub<IMetadataClient>, IMetadataServer
+    {
+        private readonly IDatabaseFactory databaseFactory;
+
+        public MetadataHub(IDatabaseFactory databaseFactory)
+        {
+            this.databaseFactory = databaseFactory;
+        }
+
+        public async Task<BeatmapUpdates> GetChangesSince(int queueId)
+        {
+            using (var db = databaseFactory.GetInstance())
+                return await db.GetUpdatedBeatmapSets(queueId);
+        }
+    }
+}
