+++ osu.Server.Spectator/Hubs/Multiplayer/IMultiplayerHubContext.cs
@@ -8,7 +8,7 @@ using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Server.Spectator.Entities;
 
-namespace osu.Server.Spectator.Hubs
+namespace osu.Server.Spectator.Hubs.Multiplayer
 {
     /// <summary>
     /// Allows communication with multiplayer clients from potentially outside of a direct <see cref="MultiplayerHub"/> context.
