+++ osu.Server.Spectator/Hubs/Multiplayer/MultiplayerQueue.cs
@@ -14,7 +14,7 @@ using osu.Server.Spectator.Database;
 using osu.Server.Spectator.Database.Models;
 using osu.Server.Spectator.Extensions;
 
-namespace osu.Server.Spectator.Hubs
+namespace osu.Server.Spectator.Hubs.Multiplayer
 {
     public class MultiplayerQueue
     {
