+++ osu.Server.Spectator/Startup.cs
@@ -16,6 +16,9 @@ using osu.Game.Online;
 using osu.Server.Spectator.Authentication;
 using osu.Server.Spectator.Extensions;
 using osu.Server.Spectator.Hubs;
+using osu.Server.Spectator.Hubs.Metadata;
+using osu.Server.Spectator.Hubs.Multiplayer;
+using osu.Server.Spectator.Hubs.Spectator;
 
 namespace osu.Server.Spectator
 {
